+++ src/components/EditMarkerDrawer.js
@@ -11,6 +11,11 @@ import { makeStyles } from '@material-ui/core/styles'
 
 import { LatLon } from 'geodesy/mgrs'
 
+//----------------------------------------------------------------//
+// Functions
+//----------------------------------------------------------------//
+import { submitCoordInput } from '../functions/submitCoordInput'
+
 const drawerWidth = 240
 
 const useStyles = makeStyles(theme => ({
@@ -46,53 +51,190 @@ export default (props) => {
   const container = props.window !== undefined ? () => window().document.body : undefined
 
   // Every Marker has these
-  const [position, setPosition] = React.useState('')
   const [title, setTitle] = React.useState('')
+  const [mgrs, setMgrs] = React.useState('')
+  const [lat, setLat] = React.useState('')
+  const [lng, setLng] = React.useState('')
+  const [elevation, setElevation] = React.useState('')
 
-  // Control the switch for 9-Lines and 15-Lines
-  const [_9Line, set9Line] = React.useState(false)
-  const [_15Line, set15Line] = React.useState(false)
+  // Control the switch for LatLng, 9-Lines, and 15-Lines, and threat fills
+  const [latlng, setLatlng] = React.useState(false)
+  const [cas, setCas] = React.useState(false)
+  const [csar, setCsar] = React.useState(false)
+  const [fill, setFill] = React.useState(false)
 
   // Control 9-Line TextFields
-  const [_9LineLabel, set9LineLabel] = React.useState('')
-  const [_9LineTypeMethod, set9LineTypeMethod] = React.useState('')
-  const [_9LineIpHdgDistance, set9LineIpHdgDistance] = React.useState('')
-  const [_9LineElevation, set9LineElevation] = React.useState('')
-  const [_9LineDescription, set9LineDescription] = React.useState('')
-  const [_9LineLocation, set9LineLocation] = React.useState('')
-  const [_9LineMark, set9LineMark] = React.useState('')
-  const [_9LineFriendlies, set9LineFriendlies] = React.useState('')
-  const [_9LineEgress, set9LineEgress] = React.useState('')
-  const [_9LineRemarks, set9LineRemarks] = React.useState('')
+  const [casLabel, setCasLabel] = React.useState('')
+  const [casTypeMethod, setCasTypeMethod] = React.useState('')
+  const [casIpHdgDistance, setCasIpHdgDistance] = React.useState('')
+  const [casElevation, setCasElevation] = React.useState('')
+  const [casDescription, setCasDescription] = React.useState('')
+  const [casLocation, setCasLocation] = React.useState('')
+  const [casMark, setCasMark] = React.useState('')
+  const [casFriendlies, setCasFriendlies] = React.useState('')
+  const [casEgress, setCasEgress] = React.useState('')
+  const [casRemarks, setCasRemarks] = React.useState('')
+
+  // Control 15-Line TextFields
+  const [csarCallsign, setCsarCallsign] = React.useState('')
+  const [csarFrequency, setCsarFrequency] = React.useState('')
+  const [csarPlsHhrid, setCsarPlsHhrid] = React.useState('')
+  const [csarNumObjectives, setCsarNumObjectives] = React.useState('')
+  const [csarLocation, setCsarLocation] = React.useState('')
+  const [csarElevation, setCsarElevation] = React.useState('')
+  const [csarDateTime, setCsarDateTime] = React.useState('')
+  const [csarSource, setCsarSource] = React.useState('')
+  const [csarCondition, setCsarCondition] = React.useState('')
+  const [csarEquipment, setCsarEquipment] = React.useState('')
+  const [csarAuthentication, setCsarAuthentication] = React.useState('')
+  const [csarThreats, setCsarThreats] = React.useState('')
+  const [csarPzDescription, setCsarPzDescription] = React.useState('')
+  const [csarOscFreq, setCsarOscFreq] = React.useState('')
+  const [csarIpHdg, setCsarIpHdg] = React.useState('')
+  const [csarRescort, setCsarRescort] = React.useState('')
+  const [csarGameplan, setCsarGameplan] = React.useState('')
+  const [csarSignal, setCsarSignal] = React.useState('')
+  const [csarEgress, setCsarEgress] = React.useState('')
 
   React.useEffect(() => {
     if (props.marker !== null) {
       setTitle(props.marker.title)
-      setPosition(LatLon.parse(props.marker.latlng).toUtm().toMgrs().toString())
+      setMgrs(LatLon.parse(props.marker.latlng).toUtm().toMgrs().toString())
+      setLat(props.marker.latlng.lat.toFixed(4))
+      setLng(props.marker.latlng.lng.toFixed(4))
+      setElevation(props.marker.elevation)
+      setCas(false)
+      setCsar(false)
 
       if (props.marker.layer === 'hostile' || props.marker.layer === 'threat') {
         if (props.marker.data === null) {
-          set9LineIpHdgDistance('N/A')
-          set9LineElevation(props.marker.elevation)
-          set9LineDescription(props.marker.title)
-          set9LineLocation(LatLon.parse(props.marker.latlng).toUtm().toMgrs().toString())
-          set9LineMark('None')
+          setCasIpHdgDistance('N/A')
+          setCasElevation(props.marker.elevation)
+          setCasDescription(props.marker.title)
+          setCasLocation(LatLon.parse(props.marker.latlng).toUtm().toMgrs().toString())
+          setCasMark('None')
         } else {
-          set9LineLabel(props.marker.data.label)
-          set9LineTypeMethod(props.marker.data.typeMethod)
-          set9LineIpHdgDistance(props.marker.data.ipHdgDistance)
-          set9LineElevation(props.marker.data.elevation)
-          set9LineDescription(props.marker.data.description)
-          set9LineLocation(props.marker.data.location)
-          set9LineMark(props.marker.data.mark)
-          set9LineFriendlies(props.marker.data.friendlies)
-          set9LineEgress(props.marker.data.egress)
-          set9LineRemarks(props.marker.data.remarks)
+          setCas(true)
+          setCasLabel(props.marker.data.label)
+          setCasTypeMethod(props.marker.data.typeMethod)
+          setCasIpHdgDistance(props.marker.data.ipHdgDistance)
+          setCasElevation(props.marker.data.elevation)
+          setCasDescription(props.marker.data.description)
+          setCasLocation(props.marker.data.location)
+          setCasMark(props.marker.data.mark)
+          setCasFriendlies(props.marker.data.friendlies)
+          setCasEgress(props.marker.data.egress)
+          setCasRemarks(props.marker.data.remarks)
+        }
+      } else if (props.marker.layer === 'survivor') {
+        if (props.marker.data === null) {
+          setCsarCallsign(props.marker.title)
+          setCsarFrequency('A')
+          setCsarNumObjectives('1')
+          setCsarLocation(LatLon.parse(props.marker.latlng).toUtm().toMgrs().toString())
+          setCsarElevation(props.marker.elevation)
+          setCsarSource('CSEL')
+          setCsarCondition('Ambulatory')
+          setCsarEquipment('Standard')
+        } else {
+          setCsar(true)
+          setCsarCallsign(props.marker.data.callsign)
+          setCsarFrequency(props.marker.data.frequency)
+          setCsarPlsHhrid(props.marker.data.plsHhrid)
+          setCsarNumObjectives(props.marker.data.numObjectives)
+          setCsarLocation(props.marker.data.location)
+          setCsarElevation(props.marker.data.elevation)
+          setCsarDateTime(props.marker.data.dateTime)
+          setCsarSource(props.marker.data.source)
+          setCsarCondition(props.marker.data.condition)
+          setCsarEquipment(props.marker.data.equipment)
+          setCsarAuthentication(props.marker.data.authentication)
+          setCsarThreats(props.marker.data.threats)
+          setCsarPzDescription(props.marker.data.pzDescription)
+          setCsarOscFreq(props.marker.data.oscFreq)
+          setCsarIpHdg(props.marker.data.ipHdg)
+          setCsarRescort(props.marker.data.rescort)
+          setCsarGameplan(props.marker.data.gameplan)
+          setCsarSignal(props.marker.data.signal)
+          setCsarEgress(props.marker.data.egress)
         }
       }
-
     }
   }, [props.marker])
+
+  const pullElevation = () => {
+    setElevation('Pending')
+
+    let target = false
+    if (latlng) {
+      target = submitCoordInput(lat + ' ' + lng)
+    } else {
+      target = submitCoordInput(mgrs)
+    }
+
+    // Get the elevation
+    fetch(`https://nationalmap.gov/epqs/pqs.php?x=${target.lng}&y=${target.lat}&units=Feet&output=json`)
+      .then(response => response.json())
+      .then(json => (Number.parseInt(json.USGS_Elevation_Point_Query_Service.Elevation_Query.Elevation) === -1000000) ? setElevation(0) : Number.parseInt(setElevation(json.USGS_Elevation_Point_Query_Service.Elevation_Query.Elevation)))
+  }
+
+  const handleSubmit = () => {
+    let target = false
+    if (latlng) {
+      target = submitCoordInput(lat + ' ' + lng)
+    } else {
+      target = submitCoordInput(mgrs)
+    }
+
+    let payload = {
+      marker: props.marker,
+      data: null,
+      elevation: elevation,
+      latlng: target,
+      title: title,
+    }
+
+    if ((props.marker.layer === 'threat' || props.marker.layer === 'hostile') && cas) {
+      payload.data = {
+        type: '9line',
+        label: casLabel,
+        typeMethod: casTypeMethod,
+        ipHdgDistance: casIpHdgDistance,
+        elevation: casElevation,
+        description: casDescription,
+        location: casLocation,
+        mark: casMark,
+        friendlies: casFriendlies,
+        egress: casEgress,
+        remarks: casRemarks,
+      }
+    } else if (props.marker.layer === 'survivor' && csar) {
+      payload.data = {
+        type: '15line',
+        callsign: csarCallsign,
+        frequency: csarFrequency,
+        plsHhrid: csarPlsHhrid,
+        numObjectives: csarNumObjectives,
+        location: csarLocation,
+        elevation: csarElevation,
+        dateTime: csarDateTime,
+        source: csarSource,
+        condition: csarCondition,
+        equipment: csarEquipment,
+        authentication: csarAuthentication,
+        threats: csarThreats,
+        pzDescription: csarPzDescription,
+        oscFreq: csarOscFreq,
+        ipHdg: csarIpHdg,
+        rescort: csarRescort,
+        gameplan: csarGameplan,
+        signal: csarSignal,
+        egress: csarEgress,
+      }
+    }
+
+    props.submit('edit', payload)
+  }
 
   return (
     <nav
@@ -114,12 +256,63 @@ export default (props) => {
           variant='outlined'
           value={title}
         />
+        <Grid
+          container
+          direction='row'
+          justify='center'
+        >
+          <FormGroup>
+            <FormControlLabel
+              control={
+                <Switch
+                  checked={latlng}
+                  color='primary'
+                  name='Lat Long'
+                  onChange={() => setLatlng(!latlng)}
+                />
+              }
+              label='Lat Long'
+            />
+          </FormGroup>
+        </Grid>
+        {(latlng) ?
+          (
+            <React.Fragment>
+              <TextField
+                className={classes.textField}
+                label='Latitude'
+                onBlur={pullElevation}
+                onChange={event => setLat(event.target.value)}
+                variant='outlined'
+                value={lat}
+              />
+              <TextField
+                className={classes.textField}
+                label='Longitude'
+                onBlur={pullElevation}
+                onChange={event => setLng(event.target.value)}
+                variant='outlined'
+                value={lng}
+              />
+            </React.Fragment>
+          ) :
+          (
+            <TextField
+              className={classes.textField}
+              label='MGRS'
+              onBlur={pullElevation}
+              onChange={event => setMgrs(event.target.value.toUpperCase())}
+              variant='outlined'
+              value={mgrs}
+            />
+          )
+        }
         <TextField
           className={classes.textField}
-          label='Position'
-          onChange={event => setPosition(event.target.value)}
+          label='Elevation'
+          onChange={event => setElevation(event.target.value)}
           variant='outlined'
-          value={position}
+          value={elevation}
         />
         {(props.marker !== null && (props.marker.layer === 'hostile' || props.marker.layer === 'threat')) ?
           (
@@ -132,10 +325,10 @@ export default (props) => {
                 <FormControlLabel
                   control={
                     <Switch
-                      checked={_9Line}
+                      checked={cas}
                       color='primary'
                       name='9 Line'
-                      onChange={() => set9Line(!_9Line)}
+                      onChange={() => setCas(!cas)}
                     />
                   }
                   label='9 Line'
@@ -145,78 +338,242 @@ export default (props) => {
           )
           : null
         }
-        {(props.marker !== null && _9Line) ?
+        {(props.marker !== null && props.marker.layer === 'survivor') ?
+          (
+            <Grid
+              container
+              direction='row'
+              justify='center'
+            >
+              <FormGroup>
+                <FormControlLabel
+                  control={
+                    <Switch
+                      checked={csar}
+                      color='primary'
+                      name='15 Line'
+                      onChange={() => setCsar(!csar)}
+                    />
+                  }
+                  label='15 Line'
+                />
+              </FormGroup>
+            </Grid>
+          )
+          : null
+        }
+        {(props.marker !== null && cas) ?
           (
             <React.Fragment>
               <TextField
                 className={classes.firstTextField}
                 label='Label'
-                onChange={event => set9LineLabel(event.target.value)}
+                onChange={event => setCasLabel(event.target.value)}
                 variant='outlined'
-                value={_9LineLabel}
+                value={casLabel}
               />
               <TextField
                 className={classes.textField}
                 label='Type / Method'
-                onChange={event => set9LineTypeMethod(event.target.value)}
+                onChange={event => setCasTypeMethod(event.target.value)}
                 variant='outlined'
-                value={_9LineTypeMethod}
+                value={casTypeMethod}
               />
               <TextField
                 className={classes.textField}
                 label='IP / HDG / Distance'
-                onChange={event => set9LineIpHdgDistance(event.target.value)}
+                onChange={event => setCasIpHdgDistance(event.target.value)}
                 variant='outlined'
-                value={_9LineIpHdgDistance}
+                value={casIpHdgDistance}
               />
               <TextField
                 className={classes.textField}
                 label='Elevation'
-                onChange={event => set9LineElevation(event.target.value)}
+                onChange={event => setCasElevation(event.target.value)}
                 variant='outlined'
-                value={_9LineElevation}
+                value={casElevation}
               />
               <TextField
                 className={classes.textField}
                 label='Description'
-                onChange={event => set9LineDescription(event.target.value)}
+                onChange={event => setCasDescription(event.target.value)}
                 variant='outlined'
-                value={_9LineDescription}
+                value={casDescription}
               />
               <TextField
                 className={classes.textField}
                 label='Location'
-                onChange={event => set9LineLocation(event.target.value)}
+                onChange={event => setCasLocation(event.target.value)}
                 variant='outlined'
-                value={_9LineLocation}
+                value={casLocation}
               />
               <TextField
                 className={classes.textField}
                 label='Mark'
-                onChange={event => set9LineMark(event.target.value)}
+                onChange={event => setCasMark(event.target.value)}
                 variant='outlined'
-                value={_9LineMark}
+                value={casMark}
               />
               <TextField
                 className={classes.textField}
                 label='Friendlies'
-                onChange={event => set9LineFriendlies(event.target.value)}
+                onChange={event => setCasFriendlies(event.target.value)}
                 variant='outlined'
-                value={_9LineFriendlies}
+                value={casFriendlies}
               />
               <TextField
                 className={classes.textField}
                 label='Egress'
-                onChange={event => set9LineEgress(event.target.value)}
+                onChange={event => setCasEgress(event.target.value)}
                 variant='outlined'
-                value={_9LineEgress}
+                value={casEgress}
               />
               <TextField
                 className={classes.textField}
                 label='Remarks &amp; Restrictions'
-                onChange={event => set9LineRemarks(event.target.value)}
+                onChange={event => setCasRemarks(event.target.value)}
                 variant='outlined'
-                value={_9LineRemarks}
+                value={casRemarks}
+              />
+            </React.Fragment>
+          )
+          : null
+        }
+        {(props.marker !== null && csar) ?
+          (
+            <React.Fragment>
+              <TextField
+                className={classes.textField}
+                label='Callsign'
+                onChange={event => setCsarCallsign(event.target.value)}
+                variant='outlined'
+                value={csarCallsign}
+              />
+              <TextField
+                className={classes.textField}
+                label='Frequency'
+                onChange={event => setCsarFrequency(event.target.value)}
+                variant='outlined'
+                value={csarFrequency}
+              />
+              <TextField
+                className={classes.textField}
+                label='PLS/HHRID'
+                onChange={event => setCsarPlsHhrid(event.target.value)}
+                variant='outlined'
+                value={csarPlsHhrid}
+              />
+              <TextField
+                className={classes.textField}
+                label='# of Objectives'
+                onChange={event => setCsarNumObjectives(event.target.value)}
+                variant='outlined'
+                value={csarNumObjectives}
+              />
+              <TextField
+                className={classes.textField}
+                label='Location'
+                onChange={event => setCsarLocation(event.target.value)}
+                variant='outlined'
+                value={csarLocation}
+              />
+              <TextField
+                className={classes.textField}
+                label='Elevation'
+                onChange={event => setCsarElevation(event.target.value)}
+                variant='outlined'
+                value={csarElevation}
+              />
+              <TextField
+                className={classes.textField}
+                label='Date/Time(Z)'
+                onChange={event => setCsarDateTime(event.target.value)}
+                variant='outlined'
+                value={csarDateTime}
+              />
+              <TextField
+                className={classes.textField}
+                label='Source'
+                onChange={event => setCsarSource(event.target.value)}
+                variant='outlined'
+                value={csarSource}
+              />
+              <TextField
+                className={classes.textField}
+                label='Condition'
+                onChange={event => setCsarCondition(event.target.value)}
+                variant='outlined'
+                value={csarCondition}
+              />
+              <TextField
+                className={classes.textField}
+                label='Equipment'
+                onChange={event => setCsarEquipment(event.target.value)}
+                variant='outlined'
+                value={csarEquipment}
+              />
+              <TextField
+                className={classes.textField}
+                label='Authentication'
+                onChange={event => setCsarAuthentication(event.target.value)}
+                variant='outlined'
+                value={csarAuthentication}
+              />
+              <TextField
+                className={classes.textField}
+                label='Threats'
+                onChange={event => setCsarThreats(event.target.value)}
+                variant='outlined'
+                value={csarThreats}
+              />
+              <TextField
+                className={classes.textField}
+                label='PZ Description'
+                onChange={event => setCsarPzDescription(event.target.value)}
+                variant='outlined'
+                value={csarPzDescription}
+              />
+              <TextField
+                className={classes.textField}
+                label='OSC/Freq'
+                onChange={event => setCsarOscFreq(event.target.value)}
+                variant='outlined'
+                value={csarOscFreq}
+              />
+              <TextField
+                className={classes.textField}
+                label='IP/Heading'
+                onChange={event => setCsarIpHdg(event.target.value)}
+                variant='outlined'
+                value={csarIpHdg}
+              />
+              <TextField
+                className={classes.textField}
+                label='Rescort'
+                onChange={event => setCsarRescort(event.target.value)}
+                variant='outlined'
+                value={csarRescort}
+              />
+              <TextField
+                className={classes.textField}
+                label='Term Area Gameplan'
+                onChange={event => setCsarGameplan(event.target.value)}
+                variant='outlined'
+                value={csarGameplan}
+              />
+              <TextField
+                className={classes.textField}
+                label='Signal'
+                onChange={event => setCsarSignal(event.target.value)}
+                variant='outlined'
+                value={csarSignal}
+              />
+              <TextField
+                className={classes.textField}
+                label='Egress'
+                onChange={event => setCsarEgress(event.target.value)}
+                variant='outlined'
+                value={csarEgress}
               />
             </React.Fragment>
           )
@@ -230,7 +587,7 @@ export default (props) => {
           <Button
             className={classes.marginsMd}
             color='primary'
-            onClick={undefined}
+            onClick={handleSubmit}
             variant='contained'
           >
             Save Changes